# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NodeWatch API",
        "description": "Self-hosted uptime monitoring service",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/monitor/{id}/logs": {
            "get": {
                "tags": [
                    "Monitors"
                ],
                "summary": "Get probe logs for a monitor",
                "operationId": "42b9ed83a5535e30ef3b34111bb9ad6a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-04-01"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-04-30"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Probe log entries grouped by region, sorted by time ascending (default: last 24 hours)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "eu-west": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MonitorLogEntry"
                                            }
                                        },
                                        "us-east": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MonitorLogEntry"
                                            }
                                        },
                                        "ap-south": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MonitorLogEntry"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MonitorLogEntry": {
                "properties": {
                    "monitor_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "checked_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-01 10:00:00"
                    },
                    "region": {
                        "type": "string",
                        "example": "eu-west",
                        "enum": [
                            "eu-west",
                            "us-east",
                            "ap-south"
                        ]
                    },
                    "status_code": {
                        "type": "integer",
                        "example": 200
                    },
                    "response_time_ms": {
                        "type": "integer",
                        "example": 142
                    },
                    "ttfb_ms": {
                        "type": "integer",
                        "example": 87
                    },
                    "is_up": {
                        "type": "boolean",
                        "example": true
                    },
                    "error": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}